# api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ACCESS Ready API - Adaptive ELL exit prep for K-12 students
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /speech/text-to-speech:
    post:
      tags: [speech]
      operationId: synthesizeSpeech
      summary: Synthesize speech from text (Azure neural TTS)
      description: |
        Converts text into spoken audio using Azure Cognitive Services. Returns
        raw MP3 bytes to be played directly by the client.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TextToSpeechBody"
      responses:
        "200":
          description: Synthesized speech audio (audio/mpeg).
          content:
            audio/mpeg:
              schema:
                type: string
                format: binary
        "400":
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "503":
          description: Server speech synthesis is not configured.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "502":
          description: Speech synthesis failed upstream.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /speech/speech-to-text:
    post:
      tags: [speech]
      operationId: transcribeSpeech
      summary: Transcribe a short spoken-answer audio clip (Azure STT)
      description: |
        Accepts raw audio bytes (e.g. "audio/webm;codecs=opus", "audio/wav") in
        the request body and returns the recognized text.
      requestBody:
        required: true
        content:
          audio/webm:
            schema:
              type: string
              format: binary
          audio/wav:
            schema:
              type: string
              format: binary
          audio/ogg:
            schema:
              type: string
              format: binary
      responses:
        "200":
          description: Transcription result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpeechToTextResponse"
        "400":
          description: Missing audio body or Content-Type header.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "503":
          description: Server speech transcription is not configured.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "502":
          description: Speech transcription failed upstream.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /schools:
    get:
      tags: [schools]
      operationId: listSchools
      summary: List all schools
      parameters:
        - name: districtId
          in: query
          required: false
          schema:
            type: string
          description: Filter by district ID
      responses:
        "200":
          description: List of schools with teacher counts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/School"
    post:
      tags: [schools]
      operationId: createSchool
      summary: Create a school
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSchoolBody"
      responses:
        "201":
          description: Created school
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/School"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /schools/{schoolId}:
    get:
      tags: [schools]
      operationId: getSchool
      summary: Get a school by ID
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: School
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/School"
        "404":
          description: School not found
    patch:
      tags: [schools]
      operationId: updateSchool
      summary: Update a school
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSchoolBody"
      responses:
        "200":
          description: Updated school
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/School"
        "404":
          description: School not found
  /schools/{schoolId}/teachers:
    get:
      tags: [schools]
      operationId: listSchoolTeachers
      summary: List teachers assigned to a school
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Teachers in this school
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Teacher"
  /schools/{schoolId}/teachers/{teacherId}:
    put:
      tags: [schools]
      operationId: linkTeacherToSchool
      summary: Assign a teacher to a school
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: string
        - name: teacherId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Teacher linked to school
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  schoolId:
                    type: string
        "404":
          description: School or teacher not found
    delete:
      tags: [schools]
      operationId: unlinkTeacherFromSchool
      summary: Remove a teacher from a school
      parameters:
        - name: schoolId
          in: path
          required: true
          schema:
            type: string
        - name: teacherId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Teacher unlinked
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
        "404":
          description: Teacher not found
...
components:
  schemas:
...
    TextToSpeechBody:
      type: object
      properties:
        text:
          type: string
          minLength: 1
          maxLength: 4000
        voice:
          type: string
      required:
        - text
    SpeechToTextResponse:
      type: object
      properties:
        text:
          type: string
      required:
        - text
...
    Teacher:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        school:
          type: ["string", "null"]
          description: Legacy free-text school name.
        schoolId:
          type: ["string", "null"]
          description: FK to the schools table. Null if teacher is not linked to a school.
        schoolName:
          type: ["string", "null"]
          description: Resolved school name from the schools table (joined).
        districtId:
          type: ["string", "null"]
          description: District ID via the linked school. Null if no school linked.
        isSolo:
          type: boolean
          description: >
            True when the teacher is independent (no schoolId).
            Solo teachers manage billing and their own student roster.
            School-managed teachers are under a principal and cannot add students.
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - isSolo
        - createdAt
...
    School:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        state:
          type: ["string", "null"]
        schoolCode:
          type: ["string", "null"]
        districtId:
          type: ["string", "null"]
        districtName:
          type: ["string", "null"]
          description: Resolved district name (joined from districts table).
        teacherCount:
          type: integer
          description: Number of teachers linked to this school.
        createdAt:
          type: string
      required:
        - id
        - name
        - createdAt
    CreateSchoolBody:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        districtId:
          type: string
          description: UUID of the parent district. Omit for independent schools.
        state:
          type: string
        schoolCode:
          type: string
      required:
        - name
...
    ErrorEnvelope:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: string
        details:
          type: object
          additionalProperties: true
...